Split utils into files
